moving edit modal to different file
