+++ main.py
@@ -11,19 +11,7 @@ def main():
     app = Dash(external_stylesheets=[BOOTSTRAP])
     app.title = "Mortgage Tracker"
     app.layout = create_layout(app, m)
-    app.run()
-
-    """m = TotalPaymentRecord()
-    mortgage = m.calculate_payments()
-
-    fig = px.line(
-        x=mortgage.index,
-        y=mortgage[PaymentSchema.principle_at_start],
-        markers=True,
-        title=PaymentSchema.principle_at_start,
-        color=mortgage[PaymentSchema.mortgage_name],
-    )
-    fig.show()"""
+    app.run(debug=False)
 
 
 if __name__ == "__main__":
