+++ src/components/layout.py
@@ -1,6 +1,12 @@
 from dash import Dash, html
 
-from . import range_dropdown, line_chart, data_dropdown, modal
+from . import (
+    range_dropdown,
+    line_chart,
+    data_dropdown,
+    edit_mortgage_modal,
+    view_mortgage_modal,
+)
 from ..mortgage_data import TotalPaymentRecord
 
 
@@ -15,7 +21,8 @@ def create_layout(app: Dash, data: TotalPaymentRecord) -> html.Div:
                 children=[
                     range_dropdown.render(app),
                     data_dropdown.render(app),
-                    modal.render(app, data.mortgage_list),
+                    view_mortgage_modal.render(app, data.mortgage_list),
+                    edit_mortgage_modal.render(app),
                 ],
             ),
             line_chart.render(app),
